docs(upload): update doc and demo 修复 初始值无法显示问题
